commented print_r in Database/Floors. Database urls are now updated after adding or removing File. FileManager got 2 more functions, getFileExtension(file) and setter + getter for TARGET_PATH

+++ htdocs/php/Model/FileForm/FileForm.php
@@ -1,6 +1,7 @@
 <?php 
 require_once 'php/Model/interface.php';
 require_once 'php/Model/FileManager/FileManager.php';
+require_once 'php/Model/Database/Floors.php';
 
 
 class FileForm implements functions_for_model
@@ -9,32 +10,62 @@ class FileForm implements functions_for_model
     {
         $output = array("Status" => "OK");
         $output["pageName"] = "FileForm";
+        $floorList = Floors::findAll();
         $b = new FileManager();
 
         $output["files"] = $b->GetList()['list'];
-        print_r($post);
+        $floorListArray = [];
+        for ($i = 0; $i < count($floorList); ++$i)
+        {
+            // get floor info for select box
+            $tmp = [$floorList[$i]->getFloorNumber(),Buildings::find($floorList[$i]->getBuildingID())->getBuildingNumber(),$floorList[$i]->getFloorID()];
+            $floorListArray[] = $tmp;
+        }
+        $output["floors"] = $floorListArray;
+
+        // print_r($post);
         if(!isset($post["fileSend"]))
         {
             return $output;
         }
-        if( $post["fileSend"] == "1")
+        if( $post["fileSend"] == "1" && isset($post['userFile']))
         {
-            $status = $b->save($post['userFile'], $post['fileName']);
+
+            // save file
+            $newFileName = "FloorId_".$post['fileName'];
+            $status = $b->save($post['userFile'], $newFileName);
+
+            // add to database
+            Floors::find($post['fileName'])->setPhotoUrl($b->getTargetPath().$newFileName.'.'.$b->getFileExtension($post['userFile']))->save();
+
             $output["saveStatus"] = $status["Status"];
         }
-
-        if( $post["fileSend"] == "2")
+        $output["files"] = $b->GetList()['list'];
+        if( $post["fileSend"] == "2" && isset($post['fileName']))
         {
             if(in_array($post['fileName'],$output["files"]))
             {
                 $status = $b->RemoveFile($post['fileName']);
                 $output["removeStatus"] = $status["Status"];
+
+                for ($i = 0; $i < count($floorList); ++$i)
+                {
+                    if($output["removeStatus"] == "Removed")
+                    {
+                        if($floorList[$i]->getPhotoUrl() == $b->getTargetPath().$post['fileName']) 
+                        {
+                            $floorList[$i]->setPhotoUrl("empty")->save();
+                        }
+                    } 
+                }
             }
             else
             {
                 $output["removeStatus"] = "file don't exist";
             }
         }
+
+
         return $output;
     }
 }
